pyproject.toml
[project]
name = "markdown_notion"
version = "0.1.0"
description = "This package converts a markdown file into Notion blocks"
readme = "README.md"
authors = [
    { name = "vicente", email = "vicente@vectorhomes.co.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "notion-client>=2.3.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
markdown_notion = "markdown_notion:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
